refactor(user form): Changed directory location
