+++ src/views/user/RegisterUserPage.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import UserForm from '../../components/userForm/UserForm';
 import './RegisterUserPage.css';
 import Navbar from '../../components/navbar/Navbar';
 import useAuth from '../../hooks/useAuth';
 import ButtonBack from '../../components/buttons/buttonBack/ButtonBack';
+import UserForm from '../../components/user/userForm/UserForm';
 
 function RegisterUserPage() {
   const { currentUser } = useAuth();
